Split custom sidebar test validator
